// src/shared/code-errors/error-codes.enum.ts
export enum ErrorCodesEnum {
  ACN001 = 'ACN001',
  ACN002 = 'ACN002',
  ACN003 = 'ACN003',
  ACN004 = 'ACN004',
  ACN005 = 'ACN005',
  ACN006 = 'ACN006',
  ACN007 = 'ACN007',
  ACN008 = 'ACN008',
  ACN009 = 'ACN009',
  ACN010 = 'ACN010',
  ACN011 = 'ACN011',
  ACN012 = 'ACN012',
  ACN013 = 'ACN013',
  ACN014 = 'ACN014',
  ACN015 = 'ACN015',
  ACN016 = 'ACN016',
  ACN017 = 'ACN017',
  ACN018 = 'ACN018',
  ACN019 = 'ACN019',
  ACN020 = 'ACN020',
  ACN021 = 'ACN021',
  ACN022 = 'ACN022',
  ACN023 = 'ACN023',
  ACN024 = 'ACN024',
  ACN025 = 'ACN025',
}